use RunUBT.sh in Linux; fielter .rsp in Source for completionHelper.cpp
use RunUBT.sh in Linux:
In Unix (Linux) environments, use RunUBT.sh instead of
calling UBT directly, as the host environment may lack
dotnet or have an incorrect version of dotnet.

fielter .rsp in Source for completionHelper.cpp:
When building compile commands for completionHelper.cpp,
focus only on the .rsp files corresponding to relevant
source files under the Source directory. This is necessary
due to numerous source files in the Plugins directory
that interfere with the calculation of
responsePathMostEntries.